// adresbook/src/app/core/modules/openapi/model/adres.ts
/**
 * Adres API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Adres { 
    street: string;
    housenumber: string;
    zipcode?: string;
    city: string;
    id: number;
}